+++ src/Component/OpenApi2/Tests/fixtures/docker-api/expected/Validator/ServiceEndpointConstraint.php
@@ -6,6 +6,6 @@ class ServiceEndpointConstraint extends \Symfony\Component\Validator\Constraints
 {
     protected function getConstraints($options) : array
     {
-        return array(new \Symfony\Component\Validator\Constraints\Collection(array('fields' => array('Spec' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'array')), new \Docker\Api\Validator\EndpointSpecConstraint(array()))), 'Ports' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'array')))), 'VirtualIPs' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'array'))))), 'allowExtraFields' => true)));
+        return array(new \Symfony\Component\Validator\Constraints\Collection(array('fields' => array('Spec' => new \Symfony\Component\Validator\Constraints\Optional(array()), 'Ports' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'array')))), 'VirtualIPs' => new \Symfony\Component\Validator\Constraints\Optional(array(new \Symfony\Component\Validator\Constraints\Type(array('0' => 'array'))))), 'allowExtraFields' => true)));
     }
 }