+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [OpenAPI] [GH#636](https://github.com/janephp/janephp/pull/636) Fixing issue with http_query_build and QS array variables
 - [OpenAPI] [GH#660](https://github.com/janephp/janephp/pull/660) Fix non-guessed OpenApiCommon validators
 - [OpenAPI] [GH#664](https://github.com/janephp/janephp/pull/664) Fix Namespace missmatch on SubModel validation
+- [JsonSchema] [GH#668](https://github.com/janephp/janephp/pull/668) Validation - Guess SubModel only on classes, not on properties
 
 ## [7.3.1] - 2022-08-22
 ### Changed